refactor: 스웨거 설정 시큐리티 관련 추가
